+++ README.md
@@ -1,5 +1,20 @@
-# TODO
+# Commands
+```sh
+docker-compose up --build
+```
+# Environment Variables
+```sh
+# .env
+EMAIL_PASSWORD=***
+EMAIL_USER=demo@gmail.com
 
-- [X] ?date range
-- [X] Implement cache in the email client for pagination: https://chatgpt.com/share/ee01972a-549c-4435-bf83-6bd7c58eade7
-- [ ] Improve cursor logic, (include page size as part of the cursor and make page size optional)
+# These can be set in the config.py and .env
+PAGE_SIZE: int = 15
+CACHE_CAPACITY_EMAIL_ID_LIST: int = 5
+CACHE_CAPACITY_EMAIL_MODEL_LIST: int = 5
+ENVIRONMENT: Literal['local', 'dev', 'prod'] = 'local'
+```
+
+# TODO: Future Improvements
+- [ ] Filter by sender
+- [ ] Filter by subject
