+++ src/client.py
@@ -4,8 +4,8 @@ import logging
 from email.message import Message
 from typing import List, Optional, Tuple
 
-from .decorators import timed_operation
-from .imap_search_criteria import IMAPSearchCriteria
+from .utils.decorators import timed_operation
+from .utils.imap_search_criteria import IMAPSearchCriteria
 
 
 class EmailClient:
